education/requirements.txt
Django==5.1
DjangoRestFramework==3.15.2